+++ src/main/java/com/donkeys_today/server/application/diary/DiaryCalenderRetriever.java
@@ -1,0 +1,4 @@
+package com.donkeys_today.server.application.diary;
+
+public class DiaryCalenderRetriever {
+}
