+++ src/main/java/com/donkeys_today/server/application/diary/DiaryService.java
@@ -5,12 +5,12 @@ import com.donkeys_today.server.domain.diary.DiaryRepository;
 import com.donkeys_today.server.domain.diary.ReplyStatus;
 import com.donkeys_today.server.domain.reply.Reply;
 import com.donkeys_today.server.domain.reply.ReplyRepository;
-import com.donkeys_today.server.presentation.Diary.dto.DiaryCalenderResponse;
-import com.donkeys_today.server.presentation.Diary.dto.DiaryContent;
-import com.donkeys_today.server.presentation.Diary.dto.DiaryFullInfo;
-import com.donkeys_today.server.presentation.Diary.dto.DiaryListResponse;
-import com.donkeys_today.server.presentation.Diary.dto.DiaryResponse;
-import com.donkeys_today.server.presentation.Diary.dto.DiarySimpleInfo;
+import com.donkeys_today.server.presentation.Diary.dto.response.DiaryCalenderResponse;
+import com.donkeys_today.server.presentation.Diary.dto.response.DiaryContent;
+import com.donkeys_today.server.presentation.Diary.dto.response.DiaryFullInfo;
+import com.donkeys_today.server.presentation.Diary.dto.response.DiaryListResponse;
+import com.donkeys_today.server.presentation.Diary.dto.response.DiaryResponse;
+import com.donkeys_today.server.presentation.Diary.dto.response.DiarySimpleInfo;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Comparator;
