+++ src/main/java/com/donkeys_today/server/presentation/Diary/dto/response/DiaryContent.java
@@ -1,4 +1,4 @@
-package com.donkeys_today.server.presentation.Diary.dto;
+package com.donkeys_today.server.presentation.Diary.dto.response;
 
 public record DiaryContent(
         String content
