+++ src/main/java/com/donkeys_today/server/presentation/Diary/dto/response/DiaryListResponse.java
@@ -1,4 +1,4 @@
-package com.donkeys_today.server.presentation.Diary.dto;
+package com.donkeys_today.server.presentation.Diary.dto.response;
 
 import java.util.List;
